rest-api/requirements.txt
alembic==1.10.4
anyio==3.6.2
click==8.1.3
fastapi==0.95.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.2
pika==1.3.2
psycopg2-binary==2.9.9
pydantic==1.10.7
PyJWT==2.7.0
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.11
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.22.0
